+++ routes/admin.route.ts
@@ -1,6 +1,8 @@
 import express from "express";
 import {
   addProduct,
+  editProduct,
+  deleteProduct,
   fetchProducts,
   fetchProductById,
 } from "../controllers/products.controller";
@@ -8,6 +10,8 @@ import {
 const router = express.Router();
 
 router.post("/add-product", addProduct);
+router.post("/edit-product", editProduct);
+router.post("/delete-product", deleteProduct);
 
 router.get("/products", fetchProducts);